Add S3 gotcha - incomplete multi-part upload costs
Each uploaded part of a multi-part upload begins accruing
storage charges immediately, even if the upload as a whole
fails. This can really add up for certain usage patterns,
and can fly under the radar easily unless you know to
look for it.

Include quick summary of this gotcha and some links (Amazon
and third-party) for more detailed information.