Add further info about limits in Certificate Manager
This is a result of of using it for a while. Unfortunately there doesn't
appear to be a WorkMail section at the moment(pointed to by one of the
tips), but is super easy to set up provided you have some knowledge of
DNS TXT and MX records, I'll add an issue acknoledging this.